requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.1
anyio==4.9.0
apimoex==1.4.0
attrs==25.3.0
blinker==1.9.0
ccxt==4.4.90
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
et_xmlfile==2.0.0
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
gunicorn
frozenlist==1.7.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.5.0
nest-asyncio==1.6.0
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.1
propcache==0.3.2
psycopg2-binary
python-dotenv
pycares==4.9.0
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
xlrd==2.0.2
yarl==1.20.1
